requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
bidict==0.23.1
blockbuster==1.5.25
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
comm==0.2.3
cryptography==44.0.3
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
executing==2.2.1
fastapi==0.118.0
forbiddenfruit==0.1.4
frozenlist==1.7.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.32.0
grpcio==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.76
langchain-groq==0.3.8
langchain-openai==0.3.33
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.8
langgraph-api==0.4.31
langgraph-checkpoint==2.1.1
langgraph-cli==0.4.2
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.14.1
langgraph-sdk==0.2.9
langsmith==0.4.31
marshmallow==3.26.1
matplotlib-inline==0.1.7
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.3
openai==1.109.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
pytest==8.4.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.48.0
structlog==25.4.0
tenacity==9.1.2
tiktoken==0.11.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
virustotal-api==1.1.11
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
